move StringManager helper class
